// frontend/src/enums/destructiveTestKeyType.ts
export enum DestructiveTestKeyType {
  GeometryTypeId = 'destructiveTestGeometryTypeId',
  ImgLongitudinal = 'destructiveTestImgLongitudinal',
  ImgTransverse = 'destructiveTestImgTransverse',
  UserId = 'destructiveTestUserId',
  ShearStrength = 'destructiveTestShearStrength',
  CompressiveStrength = 'destructiveTestCompressiveStrength',
  TensileStrength = 'destructiveTestTensileStrength',
  MomentCapacity = 'destructiveTestMomentCapacity',
  ShearCapacity = 'destructiveTestShearCapacity',
  NormalCapacity = 'destructiveTestNormalCapacity',
  Density = 'destructiveTestDensity',
  YoungsModulus = 'destructiveTestYoungsModulus',
  Date = 'destructiveTestDate'
}