// frontend/src/enums/preStressStrandKeyType.ts
export enum PreStressStrandKeyType {
  PreStressForce = 'preStressStrandForce',
  PreStressSteelClass = 'preStressStrandSteelClass',
  PreStressSteelDiameter = 'preStressStrandSteelDiameter',
  PreStressAmount = 'preStressStrandAmount',
  Date = 'preStressStrandDate',
  Location = 'preStressStrandLocation',
  ManufacturerId = "preStressManufacturerId",
}